requirements.txt
django
djangorestframework
django-cors-headers 
pymongo
mongoengine
dnspython
black